ISSUE #5520 - move contexts to top level
